size_type now 64 bit

+++ sfworks/common/FreeStorage.h
@@ -54,7 +54,7 @@
 
 namespace StringPrivate {
 
-typedef unsigned int size_type;
+typedef unsigned long int size_type;
 
 template<typename E> struct Data {
     size_type   capacity_;  // length of allocated memory block
